+++ crates/jsonschema-py/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Custom retrievers for external references. [#372](https://github.com/Stranger6667/jsonschema/issues/372)
+- Added the `mask` argument to validators for hiding sensitive data in error messages. [#434](https://github.com/Stranger6667/jsonschema/issues/434)
 
 ### Changed